requirements.txt
asgiref==3.8.1
Django==4.2.21
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.3.0
inflection==0.5.1
isort==6.0.1
packaging==25.0
pip==25.1.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
